Fix document-board, add approve dialogs

+++ src/app/board-documents/board-documents.component.ts
@@ -1,12 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, Inject, OnInit} from '@angular/core';
 import {DocumentService} from '../_services/document.service';
 import {Category} from '../models/Category';
 import {UserService} from '../_services/user.service';
-import {MatDialog} from '@angular/material/dialog';
+import {MAT_DIALOG_DATA, MatDialog, MatDialogRef} from '@angular/material/dialog';
 import {ProfileEditComponent} from '../dialogs/profile-edit/profile-edit.component';
 import {AddCategoryComponent} from '../dialogs/add-category/add-category.component';
 import {AddDocumentComponent} from '../dialogs/add-document/add-document.component';
 import {Document} from '../models/Document';
+import {ApproveComponent} from '../dialogs/approve/approve.component';
+import {TokenStorageService} from '../_services/token-storage.service';
 
 export interface DragDropListItem {
   id: string;
@@ -23,10 +25,22 @@ export class BoardDocumentsComponent implements OnInit {
 
   categories: Category[];
   documents: Document[];
+  show = false;
 
-  constructor(public documentService: DocumentService, public dialog: MatDialog) { }
+  clickButton = false;
+  documentMessage = 'Вы уверены что хотите удалить документ ';
+  categoryMessage = 'Вы уверены что хотите удалить категорию ';
+
+  constructor(public documentService: DocumentService, public dialog: MatDialog,
+              private tokenStorageService: TokenStorageService) {
+  }
 
   ngOnInit(): void {
+
+    const user = this.tokenStorageService.getUser();
+
+    this.show = user.roles.includes('ROLE_ADMIN') || user.roles.includes('ROLE_MODERATOR');
+
     this.documentService.getCategories().subscribe(
       data => {
         this.categories = JSON.parse(data);
@@ -70,8 +84,60 @@ export class BoardDocumentsComponent implements OnInit {
     });
   }
 
-  deleteCategory(id: number) {
-    this.documentService.deleteCategory(id);
+  deleteCategory(id: number, name: string) {
+    this.categoryMessage += name + '?' + ' Все документы в категории будут удалены!';
+    this.clickButton = true;
+    const dialogRef = this.dialog.open(ApproveComponent, {data: this.categoryMessage});
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === 1) {
+        console.log('Delete category with id - ', id);
+        this.documentService.deleteCategory(id);
+        window.location.reload();
+      }
+    });
   }
 
+  updateCategory(category: Category) {
+    this.clickButton = true;
+    const dialogRef = this.dialog.open(UpdateCategoryComponent, {data: category});
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === 1) {
+        console.log('Update category with id - ', category.id);
+        window.location.reload();
+      }
+    });
+  }
+
+  deleteDocument(id: number, name: string) {
+    this.documentMessage += name + '?';
+    const dialogRef = this.dialog.open(ApproveComponent, {data: this.documentMessage});
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === 1) {
+        console.log('Delete document with id - ', id);
+        this.documentService.deleteDocument(id);
+        window.location.reload();
+      }
+    });
+  }
+}
+
+
+
+@Component({
+  selector: 'update-category',
+  templateUrl: './update-category.component.html',
+  styleUrls: ['./update-category.component.css']
+})
+export class UpdateCategoryComponent {
+  category: Category = this.categoryData;
+  constructor(public dialogRef: MatDialogRef<UpdateCategoryComponent>,
+              @Inject(MAT_DIALOG_DATA) public categoryData: Category,
+              public categoryService: DocumentService) { }
+  onNoClick(): void {
+    this.dialogRef.close();
+  }
+
+  updateCategory(): void {
+    this.categoryService.updateCategory(this.categoryData);
+  }
 }
